+++ src/main/java/io/electrum/billpay/api/IAccountLookupsResource.java
@@ -8,12 +8,34 @@ import javax.ws.rs.core.SecurityContext;
 import javax.ws.rs.core.UriInfo;
 
 import io.electrum.billpay.model.AccountLookupRequest;
+import io.electrum.billpay.model.PolicyLookupRequest;
+import io.electrum.billpay.model.TrafficFineLookupRequest;
 
 public interface IAccountLookupsResource {
 
    void requestAccountInfo(
          String id,
          AccountLookupRequest body,
+         SecurityContext securityContext,
+         AsyncResponse asyncResponse,
+         Request request,
+         HttpServletRequest httpServletRequest,
+         HttpHeaders httpHeaders,
+         UriInfo uriInfo);
+
+   void requestTrafficFineInfo(
+         String id,
+         TrafficFineLookupRequest body,
+         SecurityContext securityContext,
+         AsyncResponse asyncResponse,
+         Request request,
+         HttpServletRequest httpServletRequest,
+         HttpHeaders httpHeaders,
+         UriInfo uriInfo);
+
+   void requestPolicyInfo(
+         String id,
+         PolicyLookupRequest body,
          SecurityContext securityContext,
          AsyncResponse asyncResponse,
          Request request,