+++ src/main/java/io/electrum/billpay/model/PaymentResponse.java
@@ -1,13 +1,38 @@
 package io.electrum.billpay.model;
 
+import javax.validation.constraints.NotNull;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
 import io.electrum.vas.Utils;
 import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
 
 /**
  * Represents a response to a payment request
  **/
 @ApiModel(description = "Represents a response to a payment request")
 public class PaymentResponse extends BillpayResponse {
+   protected Account account = null;
+
+   /**
+    * The customer account detail
+    **/
+   public BillpayResponse account(Account account) {
+      this.account = account;
+      return this;
+   }
+
+   @ApiModelProperty(required = true, value = "The customer account detail")
+   @JsonProperty("account")
+   @NotNull
+   public Account getAccount() {
+      return account;
+   }
+
+   public void setAccount(Account account) {
+      this.account = account;
+   }
 
    @Override
    public String toString() {
@@ -20,8 +45,9 @@ public class PaymentResponse extends BillpayResponse {
       sb.append("    client: ").append(Utils.toIndentedString(client)).append("\n");
       sb.append("    settlementEntity: ").append(Utils.toIndentedString(settlementEntity)).append("\n");
       sb.append("    receiver: ").append(Utils.toIndentedString(receiver)).append("\n");
-      sb.append("    slipData: ").append(Utils.toIndentedString(slipData)).append("\n");
       sb.append("    amounts: ").append(Utils.toIndentedString(amounts)).append("\n");
+      sb.append("    customer: ").append(Utils.toIndentedString(customer)).append("\n");
+      sb.append("    account: ").append(Utils.toIndentedString(account)).append("\n");
       sb.append("    thirdPartyIdentifiers: ").append(Utils.toIndentedString(thirdPartyIdentifiers)).append("\n");
       sb.append("    slipData: ").append(Utils.toIndentedString(slipData)).append("\n");
       sb.append("    basketRef: ").append(Utils.toIndentedString(basketRef)).append("\n");