+++ src/main/java/io/electrum/billpay/model/TrafficFineLookupRequest.java
@@ -2,57 +2,58 @@ package io.electrum.billpay.model;
 
 import javax.validation.constraints.NotNull;
 
+import org.hibernate.validator.constraints.Length;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.electrum.vas.Utils;
+import io.electrum.vas.model.Transaction;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 
 /**
- * The data resulting from an account lookup request
+ * The data required to request traffic fine info
  **/
-@ApiModel(description = "The data resulting from an account lookup request")
-public class AccountLookupResponse extends BillpayResponse {
-   protected Account account = null;
+@ApiModel(description = "The data required to request traffic fine info")
+public class TrafficFineLookupRequest extends Transaction {
+
+   private String noticeNumber = null;
 
    /**
-    * The customer account detail
+    * A reference number identifying the traffic fine to the service provider.
     **/
-   public BillpayResponse account(Account account) {
-      this.account = account;
+   public TrafficFineLookupRequest noticeNumber(String noticeNumber) {
+      this.noticeNumber = noticeNumber;
       return this;
    }
 
-   @ApiModelProperty(required = true, value = "The customer account detail")
-   @JsonProperty("account")
+   @ApiModelProperty(required = true, value = "A reference number identifying the traffic fine to the service provider.")
+   @JsonProperty("noticeNumber")
    @NotNull
-   public Account getAccount() {
-      return account;
+   @Length(min = 6, max = 40)
+   public String getNoticeNumber() {
+      return noticeNumber;
    }
 
-   public void setAccount(Account account) {
-      this.account = account;
+   public void setNoticeNumber(String noticeNumber) {
+      this.noticeNumber = noticeNumber;
    }
 
    @Override
    public String toString() {
       StringBuilder sb = new StringBuilder();
-      sb.append("class AccountLookupResponse {\n");
+      sb.append("class TrafficFineLookupRequest {\n");
 
       sb.append("    id: ").append(Utils.toIndentedString(id)).append("\n");
+      sb.append("    noticeNumber: ").append(Utils.toIndentedString(noticeNumber)).append("\n");
       sb.append("    time: ").append(Utils.toIndentedString(time)).append("\n");
       sb.append("    originator: ").append(Utils.toIndentedString(originator)).append("\n");
       sb.append("    client: ").append(Utils.toIndentedString(client)).append("\n");
       sb.append("    settlementEntity: ").append(Utils.toIndentedString(settlementEntity)).append("\n");
       sb.append("    receiver: ").append(Utils.toIndentedString(receiver)).append("\n");
-      sb.append("    slipData: ").append(Utils.toIndentedString(slipData)).append("\n");
-      sb.append("    amounts: ").append(Utils.toIndentedString(amounts)).append("\n");
-      sb.append("    customer: ").append(Utils.toIndentedString(customer)).append("\n");
-      sb.append("    account: ").append(Utils.toIndentedString(account)).append("\n");
       sb.append("    thirdPartyIdentifiers: ").append(Utils.toIndentedString(thirdPartyIdentifiers)).append("\n");
+      sb.append("    slipData: ").append(Utils.toIndentedString(slipData)).append("\n");
       sb.append("    basketRef: ").append(Utils.toIndentedString(basketRef)).append("\n");
-      sb.append("    partPaymentAllowed: ").append(Utils.toIndentedString(partPaymentAllowed)).append("\n");
-      sb.append("    overPaymentAllowed: ").append(Utils.toIndentedString(overPaymentAllowed)).append("\n");
       sb.append("}");
       return sb.toString();
    }